Add device deletion
Implemented the deletion of devices. To get this to work correctly,
quite some things had to be changed.
The preferences storing devices are now shared between activities.
The NetworkTasks still ongoing just before refreshing the list,
either adding or removing devices, will now get cancelled.
The moment the adapter is refreshed new requests are issued anyway,
so this isn't really noticeable to the user.
Used intents to pass around needed values between the activities,
the adapter, and the fragments.
Added proper back navigation from the edit activity.
Updated the Viewpager to start at the proper fragment(device) and
setting the activity title correctly.

#5 #6