Add RecyclerView data binding
Enabled data binding, this is now used to update the values of the device cards. A lot of logic has now been moved to the XLM files or was redundant. This will improve the overall code base by reducing the amount of code.
Reduces timeout from 10 seconds to 5 to make the system feel more responsive.