Add full edit functionality
Added the ability to change passwords, this is done by opening a dialog.
When no password was previously set, only a new password is required.
If there was a previous password this first needs to be met.

Implemented the actual updating of the device. This required some changes in various
places. The data structure holding the NetworkTasks is now updated.
Also, devices will only be pinged when they actually changed.

To accommodate the detection of device change, the DiffUtilCallback is now updated.
Instead of comparing memory addresses, it now properly compares using the equals method in devices and their data class. #5