Add network change handling
The main screen can now properly handle the change of a network. This is achieved by using a NetworkReceiver that receives a system intent whenever the network changes. This is then handled in either the onPause/onResume or updateNetworkChange methods. Dependent on the application running in the foreground or not. Also updated the status text generation and handling to make way for potential translations.